pyproject.toml
[tool.poetry]
name = "nupe"
version = "0.1.0"
description = "Projeto da Fabrica de Software para o NUPE do IFC - Araquari"
authors = ["\"eduardo\", \"yuri\", \"luis\", \"jorge\""]

[tool.poetry.dependencies]
python = "^3.6.1"
django = "^2.2"
djangorestframework = "^3.11.0"
validate-docbr = "^1.7.0"
drf-action-permissions = "^0.1.1"
drf-common-exceptions = "^0.1.1"
django-oauth-toolkit = "^1.3.2"
django-safedelete = "^0.5.5"
django-filter = "^2.3.0"
Pillow = "^7.1.2"
django-cors-headers = "^3.4.0"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
drf-yasg = "^1.17.1"
coverage = "^5.1"
model_bakery = "^1.1.0"
sphinx = "^3.2.1"
sphinxcontrib-redoc = "^1.6.0"
sphinx_redactor_theme = "^0.0.1"
safety = "^1.9.0"
bandit = "^1.6.2"

[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
